requirements.txt
streamlit
python-dotenv
numpy
scipy
soundfile
pydub
langchain-groq
torch
torchaudio